+++ assets/index.html
@@ -11,9 +11,12 @@
 
 <body>
   <div id="memoirable"></div>
-
+<script>
+      window.___gcfg = {
+        parsetags: 'explicit'
+      };
+    </script>
   <script type="text/javascript" src="//apis.google.com/js/client.js"></script>
-
   <script type="text/javascript" src="./vendor/react/react.js"></script>
   <script type="text/javascript" src="./vendor/react/react-dom.js"></script>
