+++ lib/dtds/index.d.ts
@@ -1,3 +1,3 @@
 /// <reference path="google-apps/googleapis-browser-common.d.ts" />
-/// <reference path="google-apps/drive-v2-browser.d.ts" />
+/// <reference path="google-apps/drive-v3-browser.d.ts" />
 /// <reference path="google-apps/plus-v1-browser.d.ts" />
