+++ package.json
@@ -24,6 +24,7 @@
   "homepage": "https://github.com/arnabdas/memoirable#readme",
   "dependencies": {
     "brace": "^0.9.0",
+    "express-http-proxy": "^0.10.1",
     "fbemitter": "2.0.2",
     "flux": "2.1.1",
     "immutable": "3.8.1",
