+++ src/app/components/authheader.tsx
@@ -2,17 +2,15 @@
 
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-
+import { browserHistory, IndexLink, Link } from 'react-router';
 import { AuthActionTypes, ProviderTypes } from '../actions/types';
 import * as AuthActions from '../actions/authActions';
-import MarkDown from '../components/markdown';
-import AuthHeader from '../components/authheader';
 import GAuthStore from '../stores/gAuthStore';
 
-import { IAuth } from '../interfaces/auth';
 
-export class Dashboard extends React.Component<{}, IAuth> {
+export default class AuthHeader extends React.Component<{}, any> {
   _listenerToken: FBEmitter.EventSubscription;
+  toggledClass: string = 'hide-header-menu';
   constructor() {
     super();
     GAuthStore.cleanState();
@@ -21,7 +19,6 @@ export class Dashboard extends React.Component<{}, IAuth> {
 
   componentDidMount() {
     this._listenerToken = GAuthStore.addChangeListener(AuthActionTypes.AUTH_GET_PROFILE, this._setStateFromStores.bind(this));
-
     AuthActions.updateProfileInfo({ provider: ProviderTypes.GOOGLE });
   }
 
@@ -33,12 +30,26 @@ export class Dashboard extends React.Component<{}, IAuth> {
     this.setState(GAuthStore.getState());
   }
 
+  saveToDrive(drive){
+   GAuthStore._saveToGoogleDrive("");
+  }
+
+  toggleClass(){
+    this.toggledClass =  this.toggledClass === 'hide-header-menu' ? 'show-header-menu' : 'hide-header-menu' ;
+    this.setState({'name':'button-state-changed'});
+  }
+
   render() {
     return (
-      <div className="row">
-        <AuthHeader></AuthHeader>
-        <MarkDown></MarkDown>
-      </div>
+      <header>
+        <nav>
+          <div>{this.state.displayName}</div>
+          <div onClick={this.toggleClass.bind(this)}> Save
+            <button id="savetodrive" className={this.toggledClass} type="button" onClick={this.saveToDrive.bind(this,'google')}>Save to google Drive</button>
+          </div>
+        </nav>
+      </header>
     );
   }
 }
+
