+++ src/app/index.tsx
@@ -5,7 +5,6 @@ import * as ReactDOM from 'react-dom';
 import { Router, Route, browserHistory, IndexRoute, Link } from 'react-router';
 
 import { MemoirableApp } from './components/memoirable';
-import { Markdown } from './components/markdown';
 import { Home } from './views/home';
 import { Dashboard } from './views/dashboard';
 
@@ -15,7 +14,6 @@ ReactDOM.render(
       <Route path="/" component={MemoirableApp}>
         <IndexRoute component={Home}/>
         <Route path="/dashboard" component={Dashboard}/>
-        <Route path="/markdown" component={Markdown}/>
       </Route>
     </Router>
   ),
