+++ src/sass/main.scss
@@ -113,3 +113,11 @@ button {
     }
   }
 }
+
+.hide-header-menu {
+  display: none;
+}
+
+.show-header-menu {
+  display: block;
+}
